+++ Project-Aurora/Project-Aurora/Modules/LightningStateManagerModule.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
+using Aurora.Modules.GameStateListen;
 using Aurora.Profiles;
 using Aurora.Settings;
-using Aurora.Utils;
 using Lombok.NET;
 
 namespace Aurora.Modules;
