+++ Project-Aurora/Project-Aurora/Modules/MediaInfoModule.cs
@@ -1,4 +1,5 @@
-﻿using Aurora.Utils;
+﻿using Aurora.Modules.Media;
+using Aurora.Utils;
 using Lombok.NET;
 
 namespace Aurora.Modules;
