+++ Project-Aurora/Project-Aurora/Profiles/Dota 2/Layers/Control_Dota2AbilityLayer.xaml.cs
@@ -15,6 +15,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using Aurora.Modules.Inputs;
 
 namespace Aurora.Profiles.Dota_2.Layers
 {
