+++ Project-Aurora/Project-Aurora/Profiles/LightingStateManager.cs
@@ -16,6 +16,7 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
+using Aurora.Modules.GameStateListen;
 
 namespace Aurora.Profiles
 {
