+++ Project-Aurora/Project-Aurora/Settings/Control_Settings.xaml.cs
@@ -11,10 +11,11 @@ using System.Windows.Media;
 using System.Windows.Navigation;
 using System.Windows.Threading;
 using Aurora.Devices.RGBNet.Config;
+using Aurora.Modules.GameStateListen;
+using Aurora.Modules.Razer;
 using Aurora.Utils;
 using Microsoft.Win32;
 using Microsoft.Win32.TaskScheduler;
-using RazerSdkHelper;
 using RazerSdkWrapper;
 using RazerSdkWrapper.Data;
 using Xceed.Wpf.Toolkit;
