+++ Project-Aurora/Project-Aurora/Settings/Window_GSIHttpDebug.xaml.cs
@@ -3,6 +3,7 @@ using System.ComponentModel;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
+using Aurora.Modules.GameStateListen;
 using Aurora.Profiles;
 using Aurora.Utils;
 using Newtonsoft.Json;
